Revert "[feature/category] 카테고리 구현"

+++ common/src/commonMain/kotlin/common/model/reseponse/paging/Pageable.kt
@@ -4,7 +4,7 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class Pageable(
-    val offset: Long,
+    val offset: Int,
     val pageNumber: Int,
     val pageSize: Int,
     val paged: Boolean,
